+++ tests/qemu_host/src/test_utils.h
@@ -18,7 +18,7 @@ int check_exported_symbols(const udynlink_module_t *p_mod, const char *slist[]);
 int check_extern_symbols(const udynlink_module_t *p_mod, const char *slist[]);
 int check_weak_symbols(const udynlink_module_t *p_mod, const char *slist[]);
 int run_test_func(const udynlink_module_t *p_mod);
-udynlink_error_t test_load_module(udynlink_module_t *p_mod, const void *base_addr, void *load_addr, uint32_t load_size, udynlink_load_mode_t load_mode);
+udynlink_error_t test_load_module(udynlink_module_t *p_mod, const void *base_addr, void *load_addr, size_t load_size, udynlink_load_mode_t load_mode);
 udynlink_error_t test_unload_module(udynlink_module_t *p_mod);
 
 void udynlink_test_add_loading_name(const char *name);
