+++ udynlink/udynlink_hash.c
@@ -13,11 +13,11 @@ void *udynlink_resolve_hashed_symbol(const udynlink_hash_table_t *table, const c
     uint32_t h = gnu_hash(name);
 
     uint32_t mask = (1u << (h % 32)) | (1u << ((h >> table->bloom_shift) % 32));
-    uint32_t bloom_idx = (h / 32) & (table->bloom_size - 1);
+    size_t bloom_idx = (h / 32) & (table->bloom_size - 1);
     if ((table->bloom[bloom_idx] & mask) != mask)
         return NULL;
 
-    uint32_t idx = table->buckets[h % table->nbuckets];
+    size_t idx = table->buckets[h % table->nbuckets];
     if (idx < table->symoffset)
         return NULL;
 
@@ -26,7 +26,7 @@ void *udynlink_resolve_hashed_symbol(const udynlink_hash_table_t *table, const c
         if ((hv | 1u) == (h | 1u)) {
             const char *sym_name = table->strtab + table->strtab_offsets[idx];
             if (strcmp(sym_name, name) == 0)
-                return (void *)(uintptr_t)table->sym_addrs[idx];
+                return (void *)table->sym_addrs[idx];
         }
         if (hv & 1u)
             break;
